edited work_here.sh to make sure a virtual environment actually exists before prompting a user to reset it

+++ Python/Projects/project1/work_here.sh
@@ -24,11 +24,13 @@ fi
 
 git pull
 
-read -p "Reset python environment? (y/n)" result
-case "$result" in
-  y|Y ) rm -fR py3 ;;
-  n|N ) true ;;
-esac
+if [ -d py3 ] ; then
+  read -p "Reset python environment? (y/n)" result
+  case "$result" in
+    y|Y ) rm -fR py3 ;;
+    n|N ) true ;;
+  esac
+fi
 
 # "If there is not a directory called py3, then do the following things:
 #    1. create a virtual environment here called py3
@@ -36,7 +38,7 @@ esac
 #        environment
 #    3. upgrade python's package installer (pip)
 #    4. install all the packages listed in requirements.txt"
-if [ ! -d ./py3 ] ; then
+if [ ! -d py3 ] ; then
   python3 -m venv py3
   source py3/bin/activate
   python3 -m pip install --upgrade pip
@@ -49,8 +51,8 @@ fi
 # and you can make changes to project1.py without worrying about losing the
 # original. It also allows you to get updated projects (with git pull) without
 # those updates overwriting your current work.
-if [ ! -f ./project1.py ] ; then
-  cp ./project1.backup.py ./project1.py
+if [ ! -f project1.py ] ; then
+  cp project1.backup.py project1.py
 fi
 
 # "activate the virtual environment"
