+++ Python/Projects/project5/work_here.sh
@@ -15,11 +15,13 @@ mkdir -p inputs outputs
 
 git pull
 
-read -p "Reset python environment? (y/n)" result
-case "$result" in
-  y|Y ) rm -fR py3 ;;
-  n|N ) true ;;
-esac
+if [ -d py3 ] ; then
+  read -p "Reset python environment? (y/n)" result
+  case "$result" in
+    y|Y ) rm -fR py3 ;;
+    n|N ) true ;;
+  esac
+fi
 
 # "If there is not a directory called py3, then do the following things:
 #    1. create a virtual environment here called py3
@@ -28,8 +30,8 @@ esac
 #    3. upgrade python's package installer (pip)
 #    4. install all the packages listed in requirements.txt"
 
-if [ ! -d ./py3 ] ; then
-  python3 -m venv ./py3
+if [ ! -d py3 ] ; then
+  python3 -m venv py3
   source py3/bin/activate
   python3 -m pip install --upgrade pip
   python3 -m pip install -r requirements.txt
@@ -41,8 +43,8 @@ fi
 # and you can make changes to project5.py without worrying about losing the
 # original. It also allows you to get updated projects (with git pull) without
 # those updates overwriting your current work.
-if [ ! -f ./project5.py ] ; then
-  cp ./project5.backup.py ./project5.py
+if [ ! -f project5.py ] ; then
+  cp project5.backup.py project5.py
 fi
 
 # "activate the virtual environment"
